requirements.txt
fastapi
uvicorn[standard]
pandas
python-docx
transformers
torch
python-multipart
